+++ packages/website/next-sitemap.config.js
@@ -1,3 +1,4 @@
+/** @type {import('next-sitemap').IConfig} */
 module.exports = {
     siteUrl: process.env.NEXT_PUBLIC_ORIGIN,
     generateRobotsTxt: true