+++ packages/virtual/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @af-utils/react-virtual-list
 
+## 0.0.10
+
+### Patch Changes
+
+-   upgraded scroll algo & types
+-   Updated dependencies
+    -   @af-utils/virtual-core@0.0.10
+
 ## 0.0.9
 
 ### Patch Changes